LPS-198264 refactor UserSearchRequestBuilder.java
moving and creating an inner class on the only usage, and deleting the other one
