+++ modules/apps/account/account-service/src/main/java/com/liferay/account/internal/retriever/AccountUserRetrieverImpl.java
@@ -5,7 +5,6 @@
 
 package com.liferay.account.internal.retriever;
 
-import com.liferay.account.internal.search.searcher.UserSearchRequestBuilder;
 import com.liferay.account.model.AccountEntry;
 import com.liferay.account.model.AccountRole;
 import com.liferay.account.retriever.AccountUserRetriever;
@@ -13,6 +12,7 @@ import com.liferay.account.service.AccountEntryLocalService;
 import com.liferay.account.service.AccountEntryUserRelLocalService;
 import com.liferay.account.service.AccountRoleLocalService;
 import com.liferay.petra.function.transform.TransformUtil;
+import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -21,18 +21,30 @@ import com.liferay.portal.kernel.search.BaseModelSearchResult;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
+import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
 import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.LinkedHashMapBuilder;
 import com.liferay.portal.kernel.util.OrderByComparator;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.search.document.Document;
 import com.liferay.portal.search.hits.SearchHits;
-import com.liferay.portal.search.searcher.SearchResponse;
+import com.liferay.portal.search.searcher.SearchRequest;
+import com.liferay.portal.search.searcher.SearchRequestBuilder;
+import com.liferay.portal.search.searcher.SearchRequestBuilderFactory;
 import com.liferay.portal.search.searcher.Searcher;
+import com.liferay.portal.search.sort.FieldSort;
+import com.liferay.portal.search.sort.SortFieldBuilder;
+import com.liferay.portal.search.sort.SortOrder;
+import com.liferay.portal.search.sort.Sorts;
 
 import java.io.Serializable;
 
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -109,37 +121,13 @@ public class AccountUserRetrieverImpl implements AccountUserRetriever {
 
 		params.put("accountEntryIds", accountEntryIds);
 
-		return _getUserBaseModelSearchResult(
-			_getSearchResponse(
-				params, cur, delta, keywords, reverse, sortField, status));
-	}
+		UserSearchRequestBuilder userSearchRequestBuilder =
+			new UserSearchRequestBuilder(
+				params, cur, delta, keywords, reverse, status, sortField);
 
-	private SearchResponse _getSearchResponse(
-		Map<String, Serializable> attributes, int cur, int delta,
-		String keywords, boolean reverse, String sortField, int status) {
-
-		return _searcher.search(
-			_userSearchRequestBuilder.attributes(
-				attributes
-			).cur(
-				cur
-			).delta(
-				delta
-			).keywords(
-				keywords
-			).reverse(
-				reverse
-			).sortField(
-				sortField
-			).status(
-				status
-			).build());
-	}
-
-	private BaseModelSearchResult<User> _getUserBaseModelSearchResult(
-		SearchResponse searchResponse) {
-
-		SearchHits searchHits = searchResponse.getSearchHits();
+		SearchHits searchHits = _searcher.search(
+			userSearchRequestBuilder.build()
+		).getSearchHits();
 
 		if (searchHits == null) {
 			if (_log.isWarnEnabled()) {
@@ -190,9 +178,124 @@ public class AccountUserRetrieverImpl implements AccountUserRetriever {
 	private Searcher _searcher;
 
 	@Reference
-	private UserLocalService _userLocalService;
+	private SearchRequestBuilderFactory _searchRequestBuilderFactory;
 
 	@Reference
-	private UserSearchRequestBuilder _userSearchRequestBuilder;
+	private SortFieldBuilder _sortFieldBuilder;
+
+	@Reference
+	private Sorts _sorts;
+
+	@Reference
+	private UserLocalService _userLocalService;
+
+	private class UserSearchRequestBuilder {
+
+		public UserSearchRequestBuilder(
+			Map<String, Serializable> attributes, int cur, int delta,
+			String keywords, boolean reverse, int status, String sortField) {
+
+			_attributes = attributes;
+			_cur = cur;
+			_delta = delta;
+			_keywords = keywords;
+			_reverse = reverse;
+			_status = status;
+			_sortField = sortField;
+		}
+
+		public SearchRequest build() {
+			SearchRequestBuilder searchRequestBuilder =
+				_searchRequestBuilderFactory.builder();
+
+			searchRequestBuilder.entryClassNames(
+				User.class.getName()
+			).withSearchContext(
+				searchContext -> {
+					boolean andSearch = false;
+
+					if (Validator.isNull(_keywords)) {
+						andSearch = true;
+					}
+					else {
+						searchContext.setKeywords(_keywords);
+					}
+
+					searchContext.setAndSearch(andSearch);
+					searchContext.setAttributes(
+						HashMapBuilder.<String, Serializable>put(
+							Field.STATUS, _status
+						).put(
+							"city", _keywords
+						).put(
+							"country", _keywords
+						).put(
+							"firstName", _keywords
+						).put(
+							"fullName", _keywords
+						).put(
+							"lastName", _keywords
+						).put(
+							"middleName", _keywords
+						).put(
+							"params", new LinkedHashMap<>()
+						).put(
+							"region", _keywords
+						).put(
+							"screenName", _keywords
+						).put(
+							"street", _keywords
+						).put(
+							"zip", _keywords
+						).putAll(
+							_attributes
+						).build());
+					searchContext.setCompanyId(
+						CompanyThreadLocal.getCompanyId());
+
+					PermissionChecker permissionChecker =
+						PermissionThreadLocal.getPermissionChecker();
+
+					if (permissionChecker != null) {
+						searchContext.setUserId(permissionChecker.getUserId());
+					}
+				}
+			).emptySearchEnabled(
+				true
+			).highlightEnabled(
+				false
+			);
+
+			if (_cur != QueryUtil.ALL_POS) {
+				searchRequestBuilder.from(_cur);
+				searchRequestBuilder.size(_delta);
+			}
+
+			if (Validator.isNotNull(_sortField)) {
+				SortOrder sortOrder = SortOrder.ASC;
+
+				if (_reverse) {
+					sortOrder = SortOrder.DESC;
+				}
+
+				FieldSort fieldSort = _sorts.field(
+					_sortFieldBuilder.getSortField(User.class, _sortField),
+					sortOrder);
+
+				searchRequestBuilder.sorts(fieldSort);
+			}
+
+			return searchRequestBuilder.build();
+		}
+
+		private Map<String, Serializable> _attributes = new HashMap<>();
+		private final int _cur;
+		private final int _delta;
+		private final String _keywords;
+		private final boolean _reverse;
+		private final String _sortField;
+		private final int _status;
+
+	}
 
 }
